LayoutTests/http/tests/security/contentSecurityPolicy/block-mixed-content-hides-warning-expected.txt

CONSOLE MESSAGE: Refused to load the script 'http://127.0.0.1:8080/security/contentSecurityPolicy/resources/alert-fail.js' because it violates the following Content Security Policy directive: "default-src 'self'".

This page should neither alert "FAIL" nor generate any mixed content warnings in the console. 
